Simul blockram: more efficient 2,2 version; untested 2,16 version
When the word lengths for both ports are the same, we can save on a lot
of conversion. This conversion was making simulation very slow. Since
the old code is still appropriate for a mixed-width version, I copied
the old code to the mixed width version that I will use later.