Bugfix for debugging: read button, make yellow yellow
What will the next commit be, bugfix for bugfix? :(

Ball is moving slowly again

As a sidenote, I've discovered that the mixed-width mapping is what
you'd intuitively expect. But I did "nextCoords" in the same cycle as
when I read the pixel, so it got combinatorially incremented to the next
coordinate lane. That's why I saw the freaky one-off mapping.