Toolbox/FIFO: Correct license copyright
Since this file was initially copied from the CλaSH source code, it
needed to retain the copyright that was on that release. The original
file didn't carry a copyright notice at the top; it relied on the
LICENSE file that was in the root of the source directory. Obviously I
do need to add it here though, and it was an oversight that this hadn't
already happened.

Note that the CλaSH source code carries the same 2-clause BSD license as
this work.